% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tables.R
\name{add_to_tables}
\alias{add_to_tables}
\title{Adds Models Results to Two Result Tables}
\usage{
add_to_tables(data, predictions, model_name, notes = "-")
}
\arguments{
\item{data}{is the full data with 'y' values and all}

\item{predictions}{is the model's prediction results}

\item{model_name}{is the name of the model that produced the predictions. For note taking purposes, only.}

\item{notes}{is additional notes for small_table}
}
\value{
a list of tables. first element is the 'big_table', second one is the 'small_table'.
    Big table contains: date, year, month, hour, model name, actual results, predicted results and WMAPE
    Small table contains: model name, WMAPE and notes
}
\description{
Adds models results to two tables. First one is big table, where every observation and prediction is visible. Second one is small table where only model name, performance and notes are visible.
}
