% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lagged.R
\name{create_lagged}
\alias{create_lagged}
\title{Creates Lagged Data}
\usage{
create_lagged(data, lags, exclude_head = NULL, exclude_tail = NULL)
}
\arguments{
\item{data}{input data}

\item{lags}{list of wanted lags. Negative values correspond to lags while positives correspond to leads.}

\item{exclude_head}{the columns of data that is not subject to lagging process (e.g. date/hour). Will be merged to left-side of the lagged data.}

\item{exclude_tail}{the columns of data that is not subject to lagging process (e.g. production). Will be merged to right-side}
}
\value{
Data with more columns as lagged versions of existing columns
}
\description{
Creates a new data frame(?) with Lag/Lead values. Also one may exclude the features(columns) from "lag process" but have them with the new data frame
}
