% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_powered.R
\name{create_powered}
\alias{create_powered}
\title{Creates Powered Data}
\usage{
create_powered(data, powers, exclude_head, exclude_tail)
}
\arguments{
\item{data}{input data}

\item{powers}{list of powers to be added. Numbers in the list should be integers > 1.}

\item{exclude_head}{the columns of data that is not subject to lagging process (e.g. date/hour). Will be merged to left-side of the lagged data.}

\item{exclude_tail}{the columns of data that is not subject to lagging process (e.g. production). Will be merged to right-side of the lagged data.}
}
\description{
Takes input data and adds powered values of the features, as new features
}
