% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_full_glmnet.R
\name{custom_full_glmnet}
\alias{custom_full_glmnet}
\title{Custom GLMNET Function for simplicity}
\usage{
custom_full_glmnet(
  train,
  trainres,
  test,
  testres,
  perf = "mae",
  output = FALSE,
  seed = 42
)
}
\arguments{
\item{train}{'x' for 'glmnet'}

\item{trainres}{'y' for 'glmnet'}

\item{test}{matrix of test data for returning predictions}

\item{testres}{observations of the test data for printing WMAPE results}

\item{perf}{type.measure for cv.glmnet. Default = 'mae'}

\item{output}{set 'TRUE' for printing WMAPE results. Default = FALSE}

\item{seed}{for setting seed. Default = 42}
}
\value{
The predictions of the glmnet model applied on the test data
}
\description{
Uses 'glmnet' package. Uses lambda.min from cv.glmnet for doing predictions on the test data. Returns the prediction results. For seeing Weighted Mean Absolute Error and Mean Absolute Error, set output to TRUE.
}
