% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_pca.R
\name{custom_pca}
\alias{custom_pca}
\title{A Custom PCA Function}
\usage{
custom_pca(train, test, n)
}
\arguments{
\item{train}{training data that PCA model will be calculated on}

\item{test}{test data for applying the PCA results on to.}

\item{n}{indicates the wanted number of Principal Components to be returned}
}
\value{
A list of first n principle components of train and test data, respectively.
}
\description{
A custom PCA function that does PCA on train data and applies the results to test. Returns first 'n' PCs of train and test results in a list.
}
