% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_autocovariance.R
\name{prepare_autocovariance}
\alias{prepare_autocovariance}
\title{Preparing for Autocovariance}
\usage{
prepare_autocovariance(data, lags)
}
\arguments{
\item{data}{input data}

\item{lags}{list of wanted lags. Negative values correspond to lags while positives correspond to leads.}
}
\description{
Prepare the Data to Lagged(?) Autocovariance
}
