% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_autocovariance.R
\name{prepare_for_autocovariance}
\alias{prepare_for_autocovariance}
\title{Preparing for Autocovariance}
\usage{
prepare_for_autocovariance(data, lags)
}
\arguments{
\item{data}{input data}

\item{lags}{list of wanted lags. Negative values correspond to lags while positives correspond to leads.}
}
\value{
data with columns as Lag/Lead
}
\description{
Prepares the Data for lagged(?) autocovariance process. Creates a new dataframe with columns as lag/lead values. (e.g. Lag1, Normal, Lead1)
}
