% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_wind_values.R
\name{scale_wind_values}
\alias{scale_wind_values}
\title{Scaling Wind Speed on Train and Test Data}
\usage{
scale_wind_values(train, test, exclude_head, exclude_tail)
}
\arguments{
\item{train}{data where the scaling coefficients are calculated and then applied}

\item{test}{data where scaling done on train is applied to}

\item{exclude_head}{the columns of data that is not subject to lagging process (e.g. date/hour). Will be merged to left-side of the lagged data.}

\item{exclude_tail}{the columns of data that is not subject to lagging process (e.g. production). Will be merged to right-side of the lagged data.}
}
\value{
list of scaled train and test data respectively
}
\description{
Scales wind speed on train data to between 0 and 1. Then applies the same coefficients to test data.
}
