% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_PCA.R
\name{temporal_PCA}
\alias{temporal_PCA}
\title{Temporal PCA}
\usage{
temporal_PCA(data, lags, features, ev, n)
}
\arguments{
\item{data}{data with lag values. Columnnames must be of format (featurename)_(lag) (e.g. ne_lag3, nw_lead1, se)}

\item{lags}{list of lags to be used. Same as create_lagged() argument}

\item{features}{list of feature names without lag/lead additions}

\item{ev}{the temporal eigenvector matrix}

\item{n}{represents how many Principal Components to be returned}
}
\value{
Time dimension reduced data. (e.g. ne_Time1, sw_Time2)
}
\description{
A function that transforms a data with lag values to a format where multiplication with eigenvectors of temporal covariance is possible.
  Then multiplicates with the eigenvectors matrix and returns the dimension-reduced data.
}
