% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmape.R
\name{wmape}
\alias{wmape}
\title{Weighted Mean Absolute Percentage Error}
\usage{
wmape(observed, predicted)
}
\arguments{
\item{observed}{the observed results}

\item{predicted}{the predicted results}
}
\value{
Calculated WMAPE result
}
\description{
Calculates Weighted Mean Absolute Percentage Error
}
